feat: retry to get the credentials in case of error
If the checkout call is done after the request has expired, safeguard
could return two errors:
"... is awaiting acknowledgement from the requesting user and the
request can no longer be used."
or
"The access request XXX has expired".

This could happen if the request already exists and just expires between
getting listed and checkout password requested.
If this happen, retry the full operation.